Create a common/icon folder + move and rename MimeIcon component
