+++ libs/search-widget/src/_video-widget/ThumbnailPlayer.svelte
@@ -1,7 +1,7 @@
 <script>
   import Thumbnail from '../widgets/row/Thumbnail.svelte';
   import { createEventDispatcher } from 'svelte';
-  import Icon from '../common/Icon.svelte';
+  import Icon from '../common/icons/Icon.svelte';
   import { fade } from 'svelte/transition';
   import Spinner from '../common/spinner/Spinner.svelte';
   import { Duration } from './transition.utils';
