+++ libs/search-widget/src/common/icons/Icon.scss
