+++ libs/search-widget/src/common/icons/MimeIcon.scss
