+++ libs/search-widget/src/viewer/Entities.svelte
@@ -6,7 +6,7 @@
   import { combineLatest, map, Observable } from 'rxjs';
   import type { EntityGroup } from '../core/models';
   import { tap } from 'rxjs/operators';
-  import Icon from '../common/Icon.svelte';
+  import Icon from '../common/icons/Icon.svelte';
 
   const allEntities = nucliaState().entities;
   const resourceEntities: Observable<EntityGroup[]> = viewerStore.resourceEntities;
